+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # v2.15.2
 
+* changed schema's file to return a file-object
+